requirements.txt
flask==2.3.3
flask-cors==3.0.10
mysql-connector-python==8.0.33
python-dotenv
